refactor/5 : view를 포함해서 구조 수정
