+++ src/main/java/baseball/Application.java
@@ -10,6 +10,5 @@ public class Application {
         baseballController.start();
 
         baseballController.count();
-
     }
 }
