--- src/main/java/baseball/Repository/ComputerRepository.java
@@ -1,18 +1,0 @@
-package baseball.Repository;
-
-import baseball.domain.Computer;
-
-public class ComputerRepository {
-
-    private final Computer computer;
-
-    public ComputerRepository(Computer computer) {
-        this.computer = computer;
-    }
-
-    public Computer getComputer() {
-        return computer;
-    }
-
-
-}
