--- src/main/java/baseball/Repository/PlayerRepository.java
@@ -1,18 +1,0 @@
-package baseball.Repository;
-
-import baseball.domain.Player;
-
-public class PlayerRepository {
-
-    private final Player player;
-
-    public PlayerRepository(Player player) {
-        this.player = player;
-    }
-
-    public Player getPlayer() {
-        return player;
-    }
-
-
-}
