+++ src/main/java/baseball/service/PlayerService.java
@@ -16,5 +16,4 @@ public class PlayerService {
         player.setInput(Console.readLine());
     }
 
-
 }
