+++ src/main/java/baseball/view/InputView.java
@@ -1,0 +1,4 @@
+package baseball.view;
+
+public class InputView {
+}
