+++ src/main/java/baseball/view/ui/Input.java
@@ -1,0 +1,5 @@
+package baseball.view.ui;
+
+public class Input {
+
+}
