+++ src/main/java/baseball/view/ui/Output.java
@@ -1,0 +1,5 @@
+package baseball.view.ui;
+
+public class Output {
+
+}
